refactor: hir::src::preproc
