+++ crates/hir/src/base_db/source_db.rs
@@ -1,27 +1,34 @@
-use preproc::source::{
-    PreprocSourceId, SourceEmittedTokenId, SourceEmittedTokenRange, SourceMacroExpansionId,
-    SourcePosition, SourcePreprocError, SourcePreprocModel, SourcePreprocUnavailable, SourceRange,
-    SourceTokenProvenance,
-};
 use rustc_hash::{FxHashMap, FxHashSet};
-use smol_str::SmolStr;
 use syntax::{
-    Compilation, ParserExpectedSyntax, PreprocessorTrace, SourceBufferOrigin, SyntaxDiagnostic,
-    SyntaxTree, SyntaxTreeBuffer, SyntaxTreeBufferIds, SyntaxTreeOptions,
+    Compilation, ParserExpectedSyntax, SyntaxDiagnostic, SyntaxTree, SyntaxTreeBuffer,
+    SyntaxTreeBufferIds,
 };
 use triomphe::Arc;
-use utils::{
-    line_index::{TextRange, TextSize},
-    path_identity::PathIdentityIndex,
-};
+use utils::{line_index::TextSize, path_identity::PathIdentityIndex};
 use vfs::{FileId, VfsPath, anchored_path::AnchoredPath};
 
 use crate::base_db::{
     compilation_plan::{self, CompilationPlan},
     diagnostics_config::{DiagnosticSource, DiagnosticsConfig},
-    project::{CompilationProfileId, Predefine, PreprocessConfig, ProjectConfig},
+    project::{CompilationProfileId, PreprocessConfig, ProjectConfig},
     source_root::{SourceRoot, SourceRootId},
 };
+
+mod preproc;
+
+pub(crate) use self::preproc::workspace_preproc_model_file_ids;
+pub use self::preproc::{
+    MappedSourcePreprocModel, PreprocExpansionDisplay, PreprocExpansionMapping,
+    PreprocExpansionSourceBuffer, PreprocManifestSource, PreprocSourceMap, PreprocSourceMapError,
+    PreprocSourceMapping, PreprocSpeculativeUniverseId, PreprocVirtualOrigin,
+    SourcePreprocContextIndex, SourcePreprocContextIndexIssue, SourcePreprocContextStatus,
+    SourcePreprocQueryError, SourcePreprocRelevantContexts, preproc_virtual_builtin_path,
+    preproc_virtual_expansion_path, preproc_virtual_predefines_path,
+    preproc_virtual_speculative_path,
+};
+#[cfg(test)]
+use self::preproc::{materialized_predefine_text, source_preproc_file_ids};
+use self::preproc::{source_preproc_context_index_for_profile, source_preproc_model};
 
 pub trait FileLoader {
     fn resolve_path(&self, path: AnchoredPath<'_>) -> Option<FileId>;
@@ -132,595 +139,6 @@ pub struct CompilationDiagnostic {
     pub diagnostic: SyntaxDiagnostic,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct MappedSourcePreprocModel {
-    pub model: SourcePreprocModel,
-    pub source_map: PreprocSourceMap,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct PreprocSourceMap {
-    entries: FxHashMap<PreprocSourceId, PreprocSourceMapping>,
-    expansion_entries: FxHashMap<SourceMacroExpansionId, PreprocExpansionMapping>,
-    predefine_sources: FxHashMap<PreprocSourceId, PreprocManifestSource>,
-    text_lengths: FxHashMap<PreprocSourceId, usize>,
-    range_offsets: FxHashMap<PreprocSourceId, usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct SourcePreprocContextIndex {
-    contexts_by_file: FxHashMap<FileId, Vec<FileId>>,
-    issues: Vec<SourcePreprocContextIndexIssue>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SourcePreprocContextIndexIssue {
-    pub model_file_id: FileId,
-    pub error: SourcePreprocQueryError,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SourcePreprocRelevantContexts {
-    pub model_file_ids: Vec<FileId>,
-    pub status: SourcePreprocContextStatus,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SourcePreprocContextStatus {
-    Complete,
-    Partial { skipped_models: usize },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PreprocSourceMapping {
-    RealFile(FileId),
-    VirtualFile { file_id: FileId, path: VfsPath, origin: PreprocVirtualOrigin },
-    VirtualDisplay { path: VfsPath, origin: PreprocVirtualOrigin },
-    Unmapped(SourcePreprocUnavailable),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PreprocExpansionMapping {
-    pub origin: PreprocVirtualOrigin,
-    pub emitted_range: SourceEmittedTokenRange,
-    pub display: PreprocExpansionDisplay,
-    pub source_buffer: PreprocExpansionSourceBuffer,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PreprocExpansionDisplay {
-    pub path: VfsPath,
-    pub text: String,
-    token_ranges: FxHashMap<SourceEmittedTokenId, TextRange>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PreprocExpansionSourceBuffer {
-    ParseStable {
-        file_id: FileId,
-        path: VfsPath,
-        text: String,
-        token_ranges: FxHashMap<SourceEmittedTokenId, TextRange>,
-    },
-    DisplayOnly {
-        path: VfsPath,
-    },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PreprocManifestSource {
-    pub file_id: FileId,
-    pub range: TextRange,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PreprocVirtualOrigin {
-    Predefines { profile: Option<CompilationProfileId> },
-    Builtin { name: SmolStr },
-    ExternalIncludeBuffer { source: PreprocSourceId },
-    Expansion { expansion: SourceMacroExpansionId },
-    Speculative { universe: PreprocSpeculativeUniverseId },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct PreprocSpeculativeUniverseId(pub u32);
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PreprocSourceMapError {
-    MissingSource {
-        source: PreprocSourceId,
-    },
-    UnmappedSource {
-        source: PreprocSourceId,
-        reason: SourcePreprocUnavailable,
-    },
-    RangeOutOfBounds {
-        source: PreprocSourceId,
-        range: TextRange,
-        mapped_range: TextRange,
-        text_len: usize,
-    },
-    MissingExpansionVirtualFile {
-        expansion: SourceMacroExpansionId,
-    },
-    MissingEmittedToken {
-        token: SourceEmittedTokenId,
-    },
-    MissingEmittedTokenRange {
-        range: SourceEmittedTokenRange,
-    },
-    DisplayOnlyVirtualSource {
-        path: VfsPath,
-        origin: PreprocVirtualOrigin,
-    },
-}
-
-impl PreprocSourceMap {
-    pub fn insert_real_file(&mut self, source: PreprocSourceId, file_id: FileId, text_len: usize) {
-        self.entries.insert(source, PreprocSourceMapping::RealFile(file_id));
-        self.predefine_sources.remove(&source);
-        self.text_lengths.insert(source, text_len);
-        self.range_offsets.insert(source, 0);
-    }
-
-    pub fn insert_virtual_file(
-        &mut self,
-        source: PreprocSourceId,
-        file_id: Option<FileId>,
-        path: VfsPath,
-        origin: PreprocVirtualOrigin,
-        text_len: usize,
-    ) {
-        self.insert_virtual_file_with_offset(source, file_id, path, origin, text_len, 0);
-    }
-
-    fn insert_virtual_file_with_offset(
-        &mut self,
-        source: PreprocSourceId,
-        file_id: Option<FileId>,
-        path: VfsPath,
-        origin: PreprocVirtualOrigin,
-        text_len: usize,
-        range_offset: usize,
-    ) {
-        let mapping = match file_id {
-            Some(file_id) => PreprocSourceMapping::VirtualFile { file_id, path, origin },
-            None => PreprocSourceMapping::VirtualDisplay { path, origin },
-        };
-        self.entries.insert(source, mapping);
-        self.predefine_sources.remove(&source);
-        self.text_lengths.insert(source, text_len);
-        self.range_offsets.insert(source, range_offset);
-    }
-
-    pub fn insert_unmapped(&mut self, source: PreprocSourceId, reason: SourcePreprocUnavailable) {
-        self.entries.insert(source, PreprocSourceMapping::Unmapped(reason));
-        self.predefine_sources.remove(&source);
-        self.text_lengths.remove(&source);
-        self.range_offsets.remove(&source);
-    }
-
-    fn insert_predefine_manifest_source(
-        &mut self,
-        source: PreprocSourceId,
-        manifest_source: PreprocManifestSource,
-    ) {
-        self.predefine_sources.insert(source, manifest_source);
-    }
-
-    pub fn get(&self, source: PreprocSourceId) -> Option<&PreprocSourceMapping> {
-        self.entries.get(&source)
-    }
-
-    pub fn predefine_manifest_source(
-        &self,
-        source: PreprocSourceId,
-    ) -> Option<PreprocManifestSource> {
-        self.predefine_sources.get(&source).copied()
-    }
-
-    pub fn insert_expansion_display_only(
-        &mut self,
-        expansion: SourceMacroExpansionId,
-        path: VfsPath,
-        display_text: String,
-        emitted_range: SourceEmittedTokenRange,
-        display_token_ranges: FxHashMap<SourceEmittedTokenId, TextRange>,
-    ) {
-        self.expansion_entries.insert(
-            expansion,
-            PreprocExpansionMapping {
-                origin: PreprocVirtualOrigin::Expansion { expansion },
-                emitted_range,
-                display: PreprocExpansionDisplay {
-                    path: path.clone(),
-                    text: display_text,
-                    token_ranges: display_token_ranges,
-                },
-                source_buffer: PreprocExpansionSourceBuffer::DisplayOnly { path },
-            },
-        );
-    }
-
-    pub fn expansion(&self, expansion: SourceMacroExpansionId) -> Option<&PreprocExpansionMapping> {
-        self.expansion_entries.get(&expansion)
-    }
-
-    pub fn expansion_display_source(
-        &self,
-        expansion: SourceMacroExpansionId,
-    ) -> Result<PreprocSourceMapping, PreprocSourceMapError> {
-        let entry = self
-            .expansion(expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        Ok(PreprocSourceMapping::VirtualDisplay {
-            path: entry.display.path.clone(),
-            origin: entry.origin.clone(),
-        })
-    }
-
-    pub fn expansion_source_buffer(
-        &self,
-        expansion: SourceMacroExpansionId,
-    ) -> Result<PreprocSourceMapping, PreprocSourceMapError> {
-        let entry = self
-            .expansion(expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        Ok(match &entry.source_buffer {
-            PreprocExpansionSourceBuffer::ParseStable { file_id, path, .. } => {
-                PreprocSourceMapping::VirtualFile {
-                    file_id: *file_id,
-                    path: path.clone(),
-                    origin: entry.origin.clone(),
-                }
-            }
-            PreprocExpansionSourceBuffer::DisplayOnly { path } => {
-                PreprocSourceMapping::VirtualDisplay {
-                    path: path.clone(),
-                    origin: entry.origin.clone(),
-                }
-            }
-        })
-    }
-
-    pub fn emitted_display_range(
-        &self,
-        expansion: SourceMacroExpansionId,
-        emitted_range: SourceEmittedTokenRange,
-    ) -> Result<TextRange, PreprocSourceMapError> {
-        let entry = self
-            .expansion(expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        emitted_range_from_token_ranges(&entry.display.token_ranges, emitted_range)
-            .ok_or(PreprocSourceMapError::MissingEmittedTokenRange { range: emitted_range })
-    }
-
-    pub fn emitted_source_buffer_range(
-        &self,
-        expansion: SourceMacroExpansionId,
-        emitted_range: SourceEmittedTokenRange,
-    ) -> Result<TextRange, PreprocSourceMapError> {
-        let entry = self
-            .expansion(expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        let PreprocExpansionSourceBuffer::ParseStable { token_ranges, .. } = &entry.source_buffer
-        else {
-            return Err(display_only_expansion_source_buffer_error(entry));
-        };
-        emitted_range_from_token_ranges(token_ranges, emitted_range)
-            .ok_or(PreprocSourceMapError::MissingEmittedTokenRange { range: emitted_range })
-    }
-
-    pub fn emitted_token_display_range(
-        &self,
-        expansion: SourceMacroExpansionId,
-        token: SourceEmittedTokenId,
-    ) -> Result<TextRange, PreprocSourceMapError> {
-        let entry = self
-            .expansion(expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        entry
-            .display
-            .token_ranges
-            .get(&token)
-            .copied()
-            .ok_or(PreprocSourceMapError::MissingEmittedToken { token })
-    }
-
-    pub fn emitted_token_source_buffer_range(
-        &self,
-        expansion: SourceMacroExpansionId,
-        token: SourceEmittedTokenId,
-    ) -> Result<TextRange, PreprocSourceMapError> {
-        let entry = self
-            .expansion(expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        let PreprocExpansionSourceBuffer::ParseStable { token_ranges, .. } = &entry.source_buffer
-        else {
-            return Err(display_only_expansion_source_buffer_error(entry));
-        };
-        token_ranges
-            .get(&token)
-            .copied()
-            .ok_or(PreprocSourceMapError::MissingEmittedToken { token })
-    }
-
-    pub fn insert_expansion_parse_stable_source_buffer(
-        &mut self,
-        expansion: SourceMacroExpansionId,
-        file_id: FileId,
-        path: VfsPath,
-        text: String,
-        token_ranges: FxHashMap<SourceEmittedTokenId, TextRange>,
-    ) -> Result<(), PreprocSourceMapError> {
-        let entry = self
-            .expansion_entries
-            .get_mut(&expansion)
-            .ok_or(PreprocSourceMapError::MissingExpansionVirtualFile { expansion })?;
-        entry.source_buffer =
-            PreprocExpansionSourceBuffer::ParseStable { file_id, path, text, token_ranges };
-        Ok(())
-    }
-
-    pub fn expansion_display_text(&self, expansion: SourceMacroExpansionId) -> Option<&str> {
-        self.expansion(expansion).map(|entry| entry.display.text.as_str())
-    }
-
-    pub fn file_id(&self, source: PreprocSourceId) -> Result<FileId, PreprocSourceMapError> {
-        match self.get(source) {
-            Some(PreprocSourceMapping::RealFile(file_id)) => Ok(*file_id),
-            Some(PreprocSourceMapping::VirtualFile { file_id, .. }) => Ok(*file_id),
-            Some(PreprocSourceMapping::VirtualDisplay { path, origin }) => {
-                Err(PreprocSourceMapError::DisplayOnlyVirtualSource {
-                    path: path.clone(),
-                    origin: origin.clone(),
-                })
-            }
-            Some(PreprocSourceMapping::Unmapped(reason)) => {
-                Err(PreprocSourceMapError::UnmappedSource { source, reason: reason.clone() })
-            }
-            None => Err(PreprocSourceMapError::MissingSource { source }),
-        }
-    }
-
-    pub fn source_positions_for_file_offset(
-        &self,
-        file_id: FileId,
-        offset: TextSize,
-    ) -> Vec<SourcePosition> {
-        let mut positions = self
-            .entries
-            .iter()
-            .filter_map(|(source, mapping)| {
-                let mapped_file_id = match mapping {
-                    PreprocSourceMapping::RealFile(mapped_file_id)
-                    | PreprocSourceMapping::VirtualFile { file_id: mapped_file_id, .. } => {
-                        *mapped_file_id
-                    }
-                    PreprocSourceMapping::VirtualDisplay { .. } => return None,
-                    PreprocSourceMapping::Unmapped(_) => return None,
-                };
-                if mapped_file_id != file_id {
-                    return None;
-                }
-
-                let range_offset = self.range_offsets.get(source).copied().unwrap_or(0);
-                let source_offset = unshift_text_size(offset, range_offset)?;
-                let text_len = self.text_lengths.get(source).copied()?;
-                (usize::from(source_offset) <= text_len)
-                    .then_some(SourcePosition { source: *source, offset: source_offset })
-            })
-            .collect::<Vec<_>>();
-        positions.sort_by_key(|position| position.source.raw());
-        positions
-    }
-
-    pub fn map_range(&self, source_range: SourceRange) -> Result<TextRange, PreprocSourceMapError> {
-        match self.get(source_range.source) {
-            Some(PreprocSourceMapping::RealFile(_))
-            | Some(PreprocSourceMapping::VirtualFile { .. })
-            | Some(PreprocSourceMapping::VirtualDisplay { .. }) => {}
-            Some(PreprocSourceMapping::Unmapped(reason)) => {
-                return Err(PreprocSourceMapError::UnmappedSource {
-                    source: source_range.source,
-                    reason: reason.clone(),
-                });
-            }
-            None => {
-                return Err(PreprocSourceMapError::MissingSource { source: source_range.source });
-            }
-        }
-
-        let range_offset = self.range_offsets.get(&source_range.source).copied().unwrap_or(0);
-        let mapped_range = shift_text_range(source_range.range, range_offset).ok_or(
-            PreprocSourceMapError::RangeOutOfBounds {
-                source: source_range.source,
-                range: source_range.range,
-                mapped_range: source_range.range,
-                text_len: usize::MAX,
-            },
-        )?;
-        let text_len = self
-            .text_lengths
-            .get(&source_range.source)
-            .copied()
-            .ok_or(PreprocSourceMapError::MissingSource { source: source_range.source })?;
-        if usize::from(mapped_range.end()) <= text_len {
-            return Ok(mapped_range);
-        }
-
-        Err(PreprocSourceMapError::RangeOutOfBounds {
-            source: source_range.source,
-            range: source_range.range,
-            mapped_range,
-            text_len,
-        })
-    }
-}
-
-impl SourcePreprocContextIndex {
-    fn push_context(&mut self, file_id: FileId, model_file_id: FileId) {
-        let contexts = self.contexts_by_file.entry(file_id).or_default();
-        if !contexts.contains(&model_file_id) {
-            contexts.push(model_file_id);
-            contexts.sort();
-        }
-    }
-
-    fn push_issue(&mut self, issue: SourcePreprocContextIndexIssue) {
-        self.issues.push(issue);
-    }
-
-    pub fn relevant_contexts(&self, file_id: FileId) -> SourcePreprocRelevantContexts {
-        SourcePreprocRelevantContexts {
-            model_file_ids: self.contexts_by_file.get(&file_id).cloned().unwrap_or_default(),
-            status: self.status(),
-        }
-    }
-
-    pub fn status(&self) -> SourcePreprocContextStatus {
-        if self.issues.is_empty() {
-            SourcePreprocContextStatus::Complete
-        } else {
-            SourcePreprocContextStatus::Partial { skipped_models: self.issues.len() }
-        }
-    }
-
-    pub fn issues(&self) -> &[SourcePreprocContextIndexIssue] {
-        &self.issues
-    }
-}
-
-fn preproc_context_file_ids(
-    mapped: &MappedSourcePreprocModel,
-    model_file_id: FileId,
-) -> Vec<FileId> {
-    let mut file_ids = Vec::new();
-    let mut seen = FxHashSet::default();
-    push_unique_file_id(&mut file_ids, &mut seen, model_file_id);
-
-    for definition in mapped.model.macro_definitions().iter() {
-        collect_context_source_range(mapped, definition.directive_range, &mut file_ids, &mut seen);
-        collect_context_source_range(mapped, definition.name_range, &mut file_ids, &mut seen);
-        if let Some(params) = &definition.params {
-            for param in params {
-                if let Some(range) = param.name_range {
-                    collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-                }
-                if let Some(range) = param.range {
-                    collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-                }
-                if let Some(default) = &param.default {
-                    for token in default {
-                        if let Some(range) = token.range {
-                            collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-                        }
-                    }
-                }
-            }
-        }
-        for token in &definition.body_tokens {
-            if let Some(range) = token.range {
-                collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-            }
-        }
-    }
-
-    for reference in mapped.model.macro_references().iter() {
-        collect_context_source_range(mapped, reference.directive_range, &mut file_ids, &mut seen);
-        collect_context_source_range(mapped, reference.name_range, &mut file_ids, &mut seen);
-    }
-
-    for call in mapped.model.macro_calls().iter() {
-        collect_context_source_range(mapped, call.call_range, &mut file_ids, &mut seen);
-        for argument in &call.arguments {
-            if let Some(range) = argument.argument_range {
-                collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-            }
-            for token in &argument.tokens {
-                if let Some(range) = token.range {
-                    collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-                }
-            }
-        }
-    }
-
-    for include in mapped.model.include_graph().directives() {
-        collect_context_source_range(mapped, include.directive_range, &mut file_ids, &mut seen);
-        if let Some(range) = include.target_range {
-            collect_context_source_range(mapped, range, &mut file_ids, &mut seen);
-        }
-        if let Some(source) = include.resolved_source {
-            collect_context_source(mapped, source, &mut file_ids, &mut seen);
-        }
-    }
-
-    for range in mapped.model.inactive_ranges() {
-        collect_context_source_range(mapped, *range, &mut file_ids, &mut seen);
-    }
-
-    for provenance in mapped.model.token_provenance().iter() {
-        match provenance {
-            SourceTokenProvenance::Source { token_range }
-            | SourceTokenProvenance::MacroBody { body_token_range: token_range, .. } => {
-                collect_context_source_range(mapped, *token_range, &mut file_ids, &mut seen);
-            }
-            SourceTokenProvenance::MacroArgument {
-                body_token_range, argument_token_range, ..
-            } => {
-                collect_context_source_range(mapped, *body_token_range, &mut file_ids, &mut seen);
-                collect_context_source_range(
-                    mapped,
-                    *argument_token_range,
-                    &mut file_ids,
-                    &mut seen,
-                );
-            }
-            SourceTokenProvenance::TokenPaste { .. }
-            | SourceTokenProvenance::Stringification { .. }
-            | SourceTokenProvenance::Builtin { .. }
-            | SourceTokenProvenance::Unavailable(_) => {}
-            SourceTokenProvenance::Predefine { source } => {
-                collect_context_source(mapped, *source, &mut file_ids, &mut seen);
-            }
-        }
-    }
-
-    file_ids.sort();
-    file_ids
-}
-
-fn collect_context_source_range(
-    mapped: &MappedSourcePreprocModel,
-    range: SourceRange,
-    file_ids: &mut Vec<FileId>,
-    seen: &mut FxHashSet<FileId>,
-) {
-    collect_context_source(mapped, range.source, file_ids, seen);
-}
-
-fn collect_context_source(
-    mapped: &MappedSourcePreprocModel,
-    source: PreprocSourceId,
-    file_ids: &mut Vec<FileId>,
-    seen: &mut FxHashSet<FileId>,
-) {
-    if let Ok(file_id) = mapped.source_map.file_id(source) {
-        push_unique_file_id(file_ids, seen, file_id);
-    }
-    if let Some(manifest_source) = mapped.source_map.predefine_manifest_source(source) {
-        push_unique_file_id(file_ids, seen, manifest_source.file_id);
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SourcePreprocQueryError {
-    UnsupportedFileKind(SourceFileKind),
-    TraceUnavailable,
-    Model(SourcePreprocError),
-    UnmappedSource { buffer_id: u32, path: String },
-}
-
 fn source_file_identity(db: &dyn SourceDb, file_id: FileId) -> SourceFileIdentity {
     let path = db.file_path(file_id).map(|path| path.to_string()).unwrap_or_default();
     let name = if path.is_empty() { "source".to_owned() } else { path.clone() };
@@ -752,491 +170,6 @@ fn insert_buffer_file_ids(
             buffer_file_ids.insert(buffer.buffer_id, file_id);
         }
     }
-}
-
-fn push_unique_file_id(file_ids: &mut Vec<FileId>, seen: &mut FxHashSet<FileId>, file_id: FileId) {
-    if seen.insert(file_id) {
-        file_ids.push(file_id);
-    }
-}
-
-fn source_preproc_file_ids(
-    db: &dyn SourceRootDb,
-    file_id: FileId,
-    profile_id: Option<CompilationProfileId>,
-    trace: &PreprocessorTrace,
-    options: &SyntaxTreeOptions,
-    preprocess: &PreprocessConfig,
-) -> Result<PreprocSourceMap, SourcePreprocQueryError> {
-    let mut source_map = PreprocSourceMap::default();
-    let path_file_ids = path_file_ids(db);
-    let root_source = PreprocSourceId::from(trace.root_buffer_id);
-    source_map.insert_real_file(root_source, file_id, db.file_text(file_id).len());
-    let include_buffer_texts = include_buffer_texts_by_path(options);
-    let predefine_sources = trace
-        .source_buffers
-        .iter()
-        .filter(|source| source.origin == SourceBufferOrigin::Predefine)
-        .map(|source| PredefineSourceBuffer {
-            source: PreprocSourceId::from(source.buffer_id),
-            text: source.text.as_deref(),
-        })
-        .collect::<Vec<_>>();
-    let predefine_map =
-        PredefineVirtualMapping::new(db, profile_id, &preprocess.predefines, predefine_sources);
-
-    for source in &trace.source_buffers {
-        let source_id = PreprocSourceId::from(source.buffer_id);
-        if source_id == root_source {
-            source_map.insert_real_file(source_id, file_id, db.file_text(file_id).len());
-            continue;
-        }
-
-        match source.origin {
-            SourceBufferOrigin::Source => {
-                if let Some(mapped_file_id) = path_file_ids.get(&source.path) {
-                    source_map.insert_real_file(
-                        source_id,
-                        mapped_file_id,
-                        db.file_text(mapped_file_id).len(),
-                    );
-                    continue;
-                }
-
-                if let Some(text) = include_buffer_texts.get(&source.path) {
-                    let path =
-                        preproc_virtual_include_buffer_path(profile_id, source_id, &source.path);
-                    let file_id = materialized_preproc_virtual_file_id(db, &path);
-                    source_map.insert_virtual_file(
-                        source_id,
-                        file_id,
-                        path,
-                        PreprocVirtualOrigin::ExternalIncludeBuffer { source: source_id },
-                        text.len(),
-                    );
-                    continue;
-                }
-
-                source_map.insert_unmapped(
-                    source_id,
-                    SourcePreprocUnavailable::DetachedSource { source: source_id },
-                );
-            }
-            SourceBufferOrigin::Predefine => {
-                if let Some(entry) = predefine_map.entry(source_id) {
-                    let manifest_source = match entry.manifest_source(db, &path_file_ids) {
-                        Ok(manifest_source) => manifest_source,
-                        Err(reason) => {
-                            source_map.insert_unmapped(source_id, reason);
-                            continue;
-                        }
-                    };
-                    source_map.insert_virtual_file_with_offset(
-                        source_id,
-                        entry.file_id,
-                        entry.path.clone(),
-                        PreprocVirtualOrigin::Predefines { profile: profile_id },
-                        entry.text_len,
-                        entry.range_offset,
-                    );
-                    if let Some(manifest_source) = manifest_source {
-                        source_map.insert_predefine_manifest_source(source_id, manifest_source);
-                    }
-                } else if let Some(reason) = predefine_map.unavailable_reason(source_id) {
-                    source_map.insert_unmapped(source_id, reason.clone());
-                } else {
-                    source_map.insert_unmapped(
-                        source_id,
-                        SourcePreprocUnavailable::DetachedSource { source: source_id },
-                    );
-                }
-            }
-        }
-    }
-
-    Ok(source_map)
-}
-
-pub fn preproc_virtual_predefines_path(profile_id: Option<CompilationProfileId>) -> VfsPath {
-    VfsPath::new_virtual_path(format!(
-        "/__vide/preproc/{}/predefines.sv",
-        profile_path_segment(profile_id)
-    ))
-}
-
-pub fn preproc_virtual_builtin_path(
-    profile_id: Option<CompilationProfileId>,
-    name: &str,
-) -> VfsPath {
-    VfsPath::new_virtual_path(format!(
-        "/__vide/preproc/{}/builtin/{}.sv",
-        profile_path_segment(profile_id),
-        sanitize_path_segment(name)
-    ))
-}
-
-pub fn preproc_virtual_expansion_path(
-    profile_id: Option<CompilationProfileId>,
-    expansion: SourceMacroExpansionId,
-) -> VfsPath {
-    VfsPath::new_virtual_path(format!(
-        "/__vide/preproc/{}/expansion/{}.sv",
-        profile_path_segment(profile_id),
-        expansion.raw()
-    ))
-}
-
-pub fn preproc_virtual_speculative_path(
-    profile_id: Option<CompilationProfileId>,
-    universe: PreprocSpeculativeUniverseId,
-    root: &str,
-) -> VfsPath {
-    VfsPath::new_virtual_path(format!(
-        "/__vide/preproc/{}/speculative/{}/{}.sv",
-        profile_path_segment(profile_id),
-        universe.0,
-        sanitize_path_segment(root)
-    ))
-}
-
-fn preproc_virtual_include_buffer_path(
-    profile_id: Option<CompilationProfileId>,
-    source_id: PreprocSourceId,
-    source_path: &str,
-) -> VfsPath {
-    VfsPath::new_virtual_path(format!(
-        "/__vide/preproc/{}/include-buffer/{}/{}.svh",
-        profile_path_segment(profile_id),
-        source_id.raw(),
-        source_basename(source_path)
-    ))
-}
-
-fn profile_path_segment(profile_id: Option<CompilationProfileId>) -> String {
-    profile_id
-        .map(|profile_id| format!("profile-{}", profile_id.0))
-        .unwrap_or_else(|| "default".to_owned())
-}
-
-fn source_basename(path: &str) -> String {
-    let name = path.rsplit(['/', '\\']).next().unwrap_or("buffer");
-    let stem = name.rsplit_once('.').map_or(name, |(stem, _)| stem);
-    sanitize_path_segment(stem)
-}
-
-fn sanitize_path_segment(input: &str) -> String {
-    let mut out = String::new();
-    for ch in input.chars() {
-        match ch {
-            'a'..='z' | 'A'..='Z' | '0'..='9' | '-' | '_' | '.' => out.push(ch),
-            _ => out.push('_'),
-        }
-    }
-    if out.is_empty() { "unnamed".to_owned() } else { out }
-}
-
-fn include_buffer_texts_by_path(options: &SyntaxTreeOptions) -> FxHashMap<String, String> {
-    options
-        .include_buffers
-        .iter()
-        .map(|buffer| (buffer.path.clone(), buffer.text.clone()))
-        .collect()
-}
-
-fn materialized_predefine_text(predefine: &str) -> String {
-    let mut definition = predefine.to_owned();
-    if let Some(index) = definition.find('=') {
-        definition.replace_range(index..index + 1, " ");
-    } else {
-        definition.push_str(" 1");
-    }
-    format!("`define {definition}\n")
-}
-
-struct PredefineVirtualMapping {
-    entries: FxHashMap<PreprocSourceId, PredefineVirtualEntry>,
-    unavailable: FxHashMap<PreprocSourceId, SourcePreprocUnavailable>,
-}
-
-struct PredefineVirtualEntry {
-    source: PreprocSourceId,
-    file_id: Option<FileId>,
-    path: VfsPath,
-    text_len: usize,
-    range_offset: usize,
-    predefine: Predefine,
-}
-
-struct PredefineSourceBuffer<'a> {
-    source: PreprocSourceId,
-    text: Option<&'a str>,
-}
-
-struct PredefineConfigEntry {
-    text: String,
-    name: SmolStr,
-    range_offset: usize,
-    predefine: Predefine,
-}
-
-impl PredefineVirtualMapping {
-    fn new(
-        db: &dyn SourceRootDb,
-        profile_id: Option<CompilationProfileId>,
-        predefines: &[Predefine],
-        sources: Vec<PredefineSourceBuffer<'_>>,
-    ) -> Self {
-        let texts = predefines
-            .iter()
-            .map(|predefine| materialized_predefine_text(predefine.as_str()))
-            .collect::<Vec<_>>();
-        let text_len = texts.iter().map(String::len).sum();
-        let path = preproc_virtual_predefines_path(profile_id);
-        let file_id = materialized_preproc_virtual_file_id(db, &path);
-        let mut range_offset = 0usize;
-        let mut configs = Vec::new();
-        for (index, predefine) in predefines.iter().enumerate() {
-            let text = &texts[index];
-            if let Some(name) = materialized_predefine_name(text) {
-                configs.push(PredefineConfigEntry {
-                    text: text.clone(),
-                    name,
-                    range_offset,
-                    predefine: predefine.clone(),
-                });
-            }
-            range_offset += text.len();
-        }
-
-        let mut configs_by_text = FxHashMap::<String, Option<usize>>::default();
-        for (index, config) in configs.iter().enumerate() {
-            let slot = configs_by_text.entry(config.text.clone()).or_insert(Some(index));
-            if *slot != Some(index) {
-                *slot = None;
-            }
-        }
-
-        let mut entries = FxHashMap::default();
-        let mut unavailable = FxHashMap::default();
-        for source in sources {
-            let Some(source_text) = source.text else {
-                unavailable.insert(
-                    source.source,
-                    SourcePreprocUnavailable::MissingPredefineSourceText { source: source.source },
-                );
-                continue;
-            };
-            let Some(config_index) = configs_by_text.get(source_text).and_then(|index| *index)
-            else {
-                unavailable.insert(
-                    source.source,
-                    SourcePreprocUnavailable::UnverifiedPredefineSource { source: source.source },
-                );
-                continue;
-            };
-            let config = &configs[config_index];
-            if materialized_predefine_name(source_text).as_ref() != Some(&config.name) {
-                unavailable.insert(
-                    source.source,
-                    SourcePreprocUnavailable::UnverifiedPredefineSource { source: source.source },
-                );
-                continue;
-            }
-            entries.insert(
-                source.source,
-                PredefineVirtualEntry {
-                    source: source.source,
-                    file_id,
-                    path: path.clone(),
-                    text_len,
-                    range_offset: config.range_offset,
-                    predefine: config.predefine.clone(),
-                },
-            );
-        }
-
-        Self { entries, unavailable }
-    }
-
-    fn entry(&self, source: PreprocSourceId) -> Option<&PredefineVirtualEntry> {
-        self.entries.get(&source)
-    }
-
-    fn unavailable_reason(&self, source: PreprocSourceId) -> Option<&SourcePreprocUnavailable> {
-        self.unavailable.get(&source)
-    }
-}
-
-impl PredefineVirtualEntry {
-    fn manifest_source(
-        &self,
-        db: &dyn SourceRootDb,
-        path_file_ids: &PathIdentityIndex<FileId>,
-    ) -> Result<Option<PreprocManifestSource>, SourcePreprocUnavailable> {
-        let Some(source) = self.predefine.source.as_ref() else {
-            return Ok(None);
-        };
-        let Some(file_id) = path_file_ids.get_path(source.path.as_path()) else {
-            return Err(SourcePreprocUnavailable::UnverifiedPredefineSource {
-                source: self.source,
-            });
-        };
-        if !manifest_predefine_source_matches(
-            db.file_text(file_id).as_ref(),
-            source.range,
-            &self.predefine,
-        ) {
-            return Err(SourcePreprocUnavailable::UnverifiedPredefineSource {
-                source: self.source,
-            });
-        }
-        Ok(Some(PreprocManifestSource { file_id, range: source.range }))
-    }
-}
-
-fn materialized_predefine_name(text: &str) -> Option<SmolStr> {
-    let rest = text.trim_start().strip_prefix("`define")?.trim_start();
-    let name =
-        rest.split(|ch: char| ch.is_whitespace() || ch == '(').next().unwrap_or_default().trim();
-    let name = name.strip_prefix('`').unwrap_or(name);
-    if name.is_empty() { None } else { Some(SmolStr::new(name)) }
-}
-
-fn manifest_predefine_source_matches(text: &str, range: TextRange, predefine: &Predefine) -> bool {
-    let start = usize::from(range.start());
-    let end = usize::from(range.end());
-    let Some(raw_source) = text.get(start..end) else {
-        return false;
-    };
-    let Some(source_definition) = unquote_manifest_predefine(raw_source) else {
-        return false;
-    };
-    source_definition == predefine.as_str()
-        && predefine_definition_name(source_definition)
-            == predefine_definition_name(predefine.as_str())
-}
-
-fn unquote_manifest_predefine(text: &str) -> Option<&str> {
-    let text = text.trim();
-    text.strip_prefix('"')
-        .and_then(|text| text.strip_suffix('"'))
-        .or_else(|| text.strip_prefix('\'').and_then(|text| text.strip_suffix('\'')))
-}
-
-fn predefine_definition_name(predefine: &str) -> Option<SmolStr> {
-    let name = predefine.split_once('=').map_or(predefine, |(name, _)| name);
-    let name = name.trim().strip_prefix('`').unwrap_or(name.trim());
-    if name.is_empty() { None } else { Some(SmolStr::new(name)) }
-}
-
-fn materialized_preproc_virtual_file_id(db: &dyn SourceRootDb, path: &VfsPath) -> Option<FileId> {
-    file_id_for_vfs_path(db, path)
-}
-
-fn file_id_for_vfs_path(db: &dyn SourceRootDb, path: &VfsPath) -> Option<FileId> {
-    for file_id in db.files().iter().copied() {
-        let source_root_id = db.source_root_id(file_id);
-        let source_root = db.source_root(source_root_id);
-        if source_root.path_for_file(&file_id) == Some(path) {
-            return Some(file_id);
-        }
-    }
-    None
-}
-
-fn shift_text_range(range: TextRange, offset: usize) -> Option<TextRange> {
-    let start = usize::from(range.start()).checked_add(offset)?;
-    let end = usize::from(range.end()).checked_add(offset)?;
-    Some(TextRange::new(
-        TextSize::from(u32::try_from(start).ok()?),
-        TextSize::from(u32::try_from(end).ok()?),
-    ))
-}
-
-fn unshift_text_size(offset: TextSize, range_offset: usize) -> Option<TextSize> {
-    let offset = usize::from(offset).checked_sub(range_offset)?;
-    Some(TextSize::from(u32::try_from(offset).ok()?))
-}
-
-fn emitted_range_from_token_ranges(
-    token_ranges: &FxHashMap<SourceEmittedTokenId, TextRange>,
-    emitted_range: SourceEmittedTokenRange,
-) -> Option<TextRange> {
-    if emitted_range.len == 0 {
-        return Some(TextRange::empty(TextSize::from(0)));
-    }
-
-    let start = emitted_range.start;
-    let end = SourceEmittedTokenId::new(start.raw().checked_add(emitted_range.len - 1)?);
-    let start_range = token_ranges.get(&start)?;
-    let end_range = token_ranges.get(&end)?;
-    Some(TextRange::new(start_range.start(), end_range.end()))
-}
-
-fn display_only_expansion_source_buffer_error(
-    entry: &PreprocExpansionMapping,
-) -> PreprocSourceMapError {
-    PreprocSourceMapError::DisplayOnlyVirtualSource {
-        path: match &entry.source_buffer {
-            PreprocExpansionSourceBuffer::ParseStable { path, .. }
-            | PreprocExpansionSourceBuffer::DisplayOnly { path } => path.clone(),
-        },
-        origin: entry.origin.clone(),
-    }
-}
-
-fn record_expansion_display_texts(
-    profile_id: Option<CompilationProfileId>,
-    model: &SourcePreprocModel,
-    source_map: &mut PreprocSourceMap,
-) {
-    for expansion in model.macro_expansions().iter() {
-        let Some((text, token_ranges)) =
-            expansion_display_text_and_ranges(model, expansion.emitted_token_range)
-        else {
-            continue;
-        };
-        let path = preproc_virtual_expansion_path(profile_id, expansion.id);
-        source_map.insert_expansion_display_only(
-            expansion.id,
-            path,
-            text,
-            expansion.emitted_token_range,
-            token_ranges,
-        );
-    }
-}
-
-fn expansion_display_text_and_ranges(
-    model: &SourcePreprocModel,
-    emitted_range: SourceEmittedTokenRange,
-) -> Option<(String, FxHashMap<SourceEmittedTokenId, TextRange>)> {
-    let mut text = String::new();
-    let mut token_ranges = FxHashMap::default();
-
-    // This is intentionally a readable display form. It is not a
-    // parse-stable SystemVerilog source buffer or source-map authority.
-    for raw in
-        emitted_range.start.raw()..emitted_range.start.raw().checked_add(emitted_range.len)?
-    {
-        let token_id = SourceEmittedTokenId::new(raw);
-        let token = model.emitted_tokens().get(token_id)?;
-        if !text.is_empty() {
-            text.push(' ');
-        }
-        let start = text.len();
-        text.push_str(token.text.as_str());
-        let end = text.len();
-        token_ranges.insert(
-            token_id,
-            TextRange::new(
-                TextSize::from(u32::try_from(start).ok()?),
-                TextSize::from(u32::try_from(end).ok()?),
-            ),
-        );
-    }
-
-    Some((text, token_ranges))
 }
 
 fn syntax_tree_options_for_file(
@@ -1478,115 +411,6 @@ fn include_buffers_for_profile(
     Arc::new(compilation_plan::include_buffers_for_plan(db, &plan))
 }
 
-pub(crate) fn workspace_preproc_model_file_ids(
-    db: &dyn SourceRootDb,
-    profile_id: Option<CompilationProfileId>,
-) -> Vec<FileId> {
-    let plan = db.compilation_plan_for_profile(profile_id);
-    let mut file_ids = FxHashSet::default();
-
-    for root in plan.roots.iter().copied() {
-        if matches!(
-            db.file_kind(root),
-            SourceFileKind::SystemVerilog | SourceFileKind::IncludeHeader
-        ) {
-            file_ids.insert(root);
-        }
-    }
-    file_ids.extend(plan.include_only.iter().copied());
-
-    for source_root_id in &plan.source_roots {
-        for candidate in db.source_root(*source_root_id).iter() {
-            if db.file_is_project_ignored(candidate) {
-                continue;
-            }
-            if matches!(db.file_kind(candidate), SourceFileKind::IncludeHeader) {
-                file_ids.insert(candidate);
-            }
-        }
-    }
-
-    for candidate in db.files().iter().copied() {
-        if db.file_is_project_ignored(candidate) {
-            continue;
-        }
-        if !matches!(db.file_kind(candidate), SourceFileKind::IncludeHeader) {
-            continue;
-        }
-        let Some(path) = db.file_path(candidate) else {
-            continue;
-        };
-        if plan.include_dirs.iter().any(|include_dir| path.starts_with(include_dir)) {
-            file_ids.insert(candidate);
-        }
-    }
-
-    let mut file_ids = file_ids.into_iter().collect::<Vec<_>>();
-    file_ids.sort();
-    file_ids
-}
-
-fn source_preproc_model(
-    db: &dyn SourceRootDb,
-    file_id: FileId,
-) -> Arc<Result<MappedSourcePreprocModel, SourcePreprocQueryError>> {
-    let file_kind = db.file_kind(file_id);
-    if !matches!(file_kind, SourceFileKind::SystemVerilog | SourceFileKind::IncludeHeader) {
-        return Arc::new(Err(SourcePreprocQueryError::UnsupportedFileKind(file_kind)));
-    }
-
-    let text = db.file_text(file_id);
-    let identity = source_file_identity(db, file_id);
-    let profile_id = db.file_compilation_profile(file_id);
-    let preprocess = db.file_preprocess_config(file_id);
-    let options = syntax_tree_options_for_file(db, file_id);
-    let Some(trace) =
-        SyntaxTree::preprocessor_trace(&text, &identity.name, &identity.path, &options)
-    else {
-        return Arc::new(Err(SourcePreprocQueryError::TraceUnavailable));
-    };
-
-    let mut source_map =
-        match source_preproc_file_ids(db, file_id, profile_id, &trace, &options, &preprocess) {
-            Ok(source_map) => source_map,
-            Err(err) => return Arc::new(Err(err)),
-        };
-    let model = match SourcePreprocModel::from_trace(trace) {
-        Ok(model) => model,
-        Err(err) => return Arc::new(Err(SourcePreprocQueryError::Model(err))),
-    };
-    record_expansion_display_texts(profile_id, &model, &mut source_map);
-
-    Arc::new(Ok(MappedSourcePreprocModel { model, source_map }))
-}
-
-fn source_preproc_context_index_for_profile(
-    db: &dyn SourceRootDb,
-    profile_id: Option<CompilationProfileId>,
-) -> Arc<SourcePreprocContextIndex> {
-    let mut index = SourcePreprocContextIndex::default();
-
-    for model_file_id in workspace_preproc_model_file_ids(db, profile_id) {
-        index.push_context(model_file_id, model_file_id);
-        let mapped = db.source_preproc_model(model_file_id);
-        match mapped.as_ref() {
-            Ok(mapped) => {
-                for file_id in preproc_context_file_ids(mapped, model_file_id) {
-                    index.push_context(file_id, model_file_id);
-                }
-            }
-            Err(error) => {
-                index.push_issue(SourcePreprocContextIndexIssue {
-                    model_file_id,
-                    error: error.clone(),
-                });
-            }
-        }
-    }
-
-    Arc::new(index)
-}
-
 fn macro_reference_index_for_profile(
     db: &dyn SourceRootDb,
     profile_id: Option<CompilationProfileId>,
@@ -1807,14 +631,20 @@ fn source_root_semantic_diagnostics(
 mod tests {
     use std::fmt;
 
+    use ::preproc::source::{
+        PreprocSourceId, SourceMacroExpansionId, SourcePreprocUnavailable, SourceRange,
+    };
     use rustc_hash::FxHashSet;
-    use syntax::{SourceBufferId, SourceBufferOrigin, SyntaxTreeOptions};
-    use utils::paths::{AbsPathBuf, Utf8PathBuf};
+    use syntax::{PreprocessorTrace, SourceBufferId, SourceBufferOrigin, SyntaxTreeOptions};
+    use utils::{
+        line_index::TextRange,
+        paths::{AbsPathBuf, Utf8PathBuf},
+    };
     use vfs::{FileSet, VfsPath};
 
     use super::*;
     use crate::base_db::{
-        project::{CompilationProfile, PredefineSource},
+        project::{CompilationProfile, Predefine, PredefineSource},
         salsa::{self, Durability},
     };
