+++ crates/hir/src/preproc/helpers.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+mod context;
+mod expansion;
+mod facts;
+mod source;
+
+pub(in crate::preproc) use self::{context::*, expansion::*, facts::*, source::*};
